+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [Unreleased]
+### Fixed
+ - Fix `wineserver -w` calls hanging when legacy Steam Runtime and background wineserver are enabled
+
 ## [1.8.2] - 2022-05-16
 ### Fixed
  - Fix Wine crash on newer Steam Runtime installations due to renamed runtime executable