+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
-## [Unreleased]
+## [1.9.0] - 2022-07-02
 ### Added
  - Add `-l/--list` command to list all games
 
@@ -216,7 +216,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - The last release of Protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
 
-[Unreleased]: https://github.com/Matoking/protontricks/compare/1.8.2...HEAD
+[Unreleased]: https://github.com/Matoking/protontricks/compare/1.9.0...HEAD
+[1.9.0]: https://github.com/Matoking/protontricks/compare/1.8.2...1.9.0
 [1.8.2]: https://github.com/Matoking/protontricks/compare/1.8.1...1.8.2
 [1.8.1]: https://github.com/Matoking/protontricks/compare/1.8.0...1.8.1
 [1.8.0]: https://github.com/Matoking/protontricks/compare/1.7.0...1.8.0